Three fixes, one refusal with an entry condition for Plan 3, and one reviewer
proposal rejected in favour of the pattern already in the file.

The GGUF type tag is now compared as an int32_t and never cast to the enum --
positive validation of what our own writer emits, not a range check against
ggml's constant. A package promising more Reference Audio clips than the runtime
reads is refused at load in both families, verified against the published
packages by reading the real files rather than the conversion scripts. WAV
chunks are sized against the file in all four test readers.

The bot review of this branch then found a genuine one-byte heap overflow in the
same function the branch had just hardened: the chunk bound refuses a length
larger than the file, but an odd data chunk's length is real -- the buffer was
short because the resize truncated. Fixed with an ASan-verified test.

The mel band guard was refused: the function it would protect is in an anonymous
namespace and cannot be reached, and a guard nobody can reach admits no test
that fails when it is deleted. Plan 3 gets an entry condition instead.

Carries the Plan 2 carry-over ledger, which opens by recording that it was
written after the plan workspace was deleted, so Plan 2's per-task deferred
minors are unrecoverable.

+++ src/synthesize.cpp
@@ -385,7 +385,7 @@ synth::ModelInfo shared_info(const synth::omnivoice::ModelInfo &        info,
     profile.min_frames_per_clip          = info.profile.min_frames_per_clip;
     profile.max_frames_per_clip          = info.profile.max_frames_per_clip;
     profile.max_total_frames             = info.profile.max_total_frames;
-    profile.max_reference_count          = static_cast<uint32_t>(info.profile.max_reference_count);
+    profile.max_reference_count          = info.profile.max_reference_count;
     // Best-effort: HParams::profile is already validated at load time
     // (weights.cpp's read_profile_contract, is_sha256_hex), so this should
     // never fail for a package that made it this far; a defect that slipped
